+++ paper/src/main/kotlin/dev/nyon/magnetic/Main.kt
@@ -5,6 +5,10 @@ import dev.nyon.konfig.config.loadConfig
 import org.bukkit.Bukkit
 import org.bukkit.NamespacedKey
 import org.bukkit.plugin.java.JavaPlugin
+import java.nio.file.Path
+import kotlin.io.path.createParentDirectories
+import kotlin.io.path.exists
+import kotlin.io.path.moveTo
 import dev.nyon.magnetic.config as internalConfig
 
 val magneticKey = NamespacedKey("magnetic", "magnetic")
@@ -16,13 +20,20 @@ class Main : JavaPlugin() {
 
     override fun onLoad() {
         INSTANCE = this
-        config(Bukkit.getPluginsFolder().toPath().resolve("magnetic.json"), 1, Config()) { _, _, _ -> null }
+        val configPath = Bukkit.getPluginsFolder().toPath().resolve("magnetic/magnetic.json")
+        moveConfigToNewPath(configPath)
+        config(configPath, 1, Config()) { _, _, _ -> null }
         internalConfig = loadConfig()
     }
 
     override fun onEnable() {
         Listeners.listenForBukkitEvents()
         Listeners.listenForModEvents()
+    }
+
+    private fun moveConfigToNewPath(newPath: Path) {
+        val oldPath = Bukkit.getPluginsFolder().toPath().resolve("magnetic.json")
+        if (oldPath.exists()) oldPath.moveTo(newPath.createParentDirectories(), overwrite = true)
     }
 }