+++ changelog.md
@@ -7,5 +7,6 @@
   - add feature for dropped leafs
 
 # Paper
+- move config from `/plugins/magnetic.json` to `/plugins/magnetic/magnetic.json`
 - resolve ClassCastException - by @Aathoss in [#41](https://github.com/btwonion/magnetic/pull/41)
 - fix statistic not being updated properly